+++ apps/cli/src/__tests__/__examples__/action-double-quote.ts
@@ -1,0 +1,13 @@
+import { createMachine } from "xstate";
+
+createMachine({
+  tsTypes: {} as import("./action-double-quote.typegen").Typegen0,
+  schema: {
+    events: {} as { type: "PING" },
+  },
+  on: {
+    PING: {
+      actions: 'show me your "tricks"',
+    },
+  },
+});