+++ apps/cli/src/__tests__/__examples__/delay-single-quote.ts
@@ -1,0 +1,14 @@
+import { createMachine } from "xstate";
+
+createMachine({
+  tsTypes: {} as import("./delay-single-quote.typegen").Typegen0,
+  initial: "a",
+  states: {
+    a: {
+      after: {
+        "just '2 minutes'": "b",
+      },
+    },
+    b: {},
+  },
+});