+++ apps/cli/src/__tests__/__examples__/event-backtick.ts
@@ -1,0 +1,13 @@
+import { createMachine } from "xstate";
+
+createMachine({
+  tsTypes: {} as import("./event-backtick.typegen").Typegen0,
+  schema: {
+    events: {} as { type: "`tick`, `tick`" },
+  },
+  on: {
+    "`tick`, `tick`": {
+      actions: "testAction",
+    },
+  },
+});