+++ packages/machine-extractor/src/options.ts
@@ -1,10 +1,15 @@
+import { ActionNode, ChooseAction } from "./actions";
 import { maybeIdentifierTo } from "./identifiers";
 import { AnyNode, BooleanLiteral } from "./scalars";
 import { maybeTsAsExpression } from "./tsAsExpression";
+import { unionType } from "./unionType";
 import { objectOf, objectTypeWithKnownKeys } from "./utils";
+import { types as t } from "@babel/core";
 
 const MachineOptionsObject = objectTypeWithKnownKeys({
-  actions: objectOf(AnyNode),
+  actions: objectOf(
+    unionType<ActionNode | { node: t.Node }>([ChooseAction, AnyNode])
+  ),
   services: objectOf(AnyNode),
   guards: objectOf(AnyNode),
   delays: objectOf(AnyNode),
@@ -12,5 +17,5 @@ const MachineOptionsObject = objectTypeWithKnownKeys({
 });
 
 export const MachineOptions = maybeTsAsExpression(
-  maybeIdentifierTo(MachineOptionsObject),
+  maybeIdentifierTo(MachineOptionsObject)
 );