+++ packages/shared/src/index.ts
@@ -1,6 +1,5 @@
 export * from "./getTransitionsFromNode";
 export * from "./introspectMachine";
-export * from "./makeInterfaceFromIntrospectionResult";
 export * from "./filterOutIgnoredMachines";
 export * from "./isCursorInPosition";
 export * from "./getRangeFromSourceLocation";