+++ apps/extension/client/CHANGELOG.md
@@ -1,5 +1,48 @@
 # stately-vscode
 
+## 1.8.1
+
+### Patch Changes
+
+- [`2241137`](https://github.com/statelyai/xstate-tools/commit/22411372fd36490c21ff1ab63cbcc779448074f4) Thanks [@mattpocock](https://github.com/mattpocock)! - Fixed a failed deploy
+
+## 1.8.0
+
+### Minor Changes
+
+- [#151](https://github.com/statelyai/xstate-tools/pull/151) [`795a057`](https://github.com/statelyai/xstate-tools/commit/795a057f73f0a38784548a1fcf055757f44d0647) Thanks [@mattpocock](https://github.com/mattpocock)! - Added parsing for the `@xstate/test` `createTestMachine` function
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @xstate/tools-shared@1.1.6
+
+## 1.7.6
+
+### Patch Changes
+
+- [#140](https://github.com/statelyai/xstate-tools/pull/140) [`e073819`](https://github.com/statelyai/xstate-tools/commit/e0738191c61290c8f5a9ecdd507e6418ab551518) Thanks [@Andarist](https://github.com/Andarist)! - Fixed an issue with not being able to parse type-only import/export specifiers.
+
+* [#130](https://github.com/statelyai/xstate-tools/pull/130) [`99e4cb5`](https://github.com/statelyai/xstate-tools/commit/99e4cb57f3590448ddbcdc85a3104d29ef0fa79c) Thanks [@mattpocock](https://github.com/mattpocock)! - Fixed a bug where actions and conditions inside `choose` inside machine options would not be found in typegen.
+
+* Updated dependencies [[`99e4cb5`](https://github.com/statelyai/xstate-tools/commit/99e4cb57f3590448ddbcdc85a3104d29ef0fa79c)]:
+  - @xstate/tools-shared@1.1.5
+
+## 1.7.5
+
+### Patch Changes
+
+- [#122](https://github.com/statelyai/xstate-tools/pull/122) [`a725a19`](https://github.com/statelyai/xstate-tools/commit/a725a19ee6d5d6ad51a42cc8b40deeb5d6ae6215) Thanks [@mattpocock](https://github.com/mattpocock)! - Fixed an issue where warnings for meta properties would still show even if the user had turned off visual editor warnings.
+
+* [#126](https://github.com/statelyai/xstate-tools/pull/126) [`c705a64`](https://github.com/statelyai/xstate-tools/commit/c705a64d95fa99046a7acd77f16b9b0dddd2e7ba) Thanks [@ericjonathan6](https://github.com/ericjonathan6)! - Fixed a bug where descriptions on transitions were not being visualised in VSCode.
+
+- [#118](https://github.com/statelyai/xstate-tools/pull/118) [`58628f1`](https://github.com/statelyai/xstate-tools/commit/58628f182faeace8e61c995172995f98c9d623af) Thanks [@Andarist](https://github.com/Andarist)! - Fixed issues with typegen generation for quotes in tags and state names.
+
+* [#128](https://github.com/statelyai/xstate-tools/pull/128) [`2d88c75`](https://github.com/statelyai/xstate-tools/commit/2d88c75bb475cd8cee74677b293aaffe124b4d3b) Thanks [@Andarist](https://github.com/Andarist)! - Fixed issues with typegen generation for quotes in actions, delays, guards, services and event types.
+
+* Updated dependencies [[`58628f1`](https://github.com/statelyai/xstate-tools/commit/58628f182faeace8e61c995172995f98c9d623af), [`2d88c75`](https://github.com/statelyai/xstate-tools/commit/2d88c75bb475cd8cee74677b293aaffe124b4d3b), [`7a7f308`](https://github.com/statelyai/xstate-tools/commit/7a7f308672540d95a1c9292a32c60d31c5208d13)]:
+  - @xstate/tools-shared@1.1.4
+
 ## 1.7.4
 
 ### Patch Changes